Fix inference of bounds for members with ptr interop types. (#470)
The compiler assert during code generation described in issue #458 is still happening.   There was a problem with inference of bounds for a member with a ptr interop type.  We were using the base expression of the member expression, not the member expression itself.

This change adds tests that check the inference of bounds for members with interop types.  It also updates the regression test case to generate an object file, instead of just doing semantic checking.  During testing, I found that checking of bounds declarations was not happening for some bounds-safe interface assignments (assignments in unchecked scopes of checked pointers to unchecked pointer lvalues that have bounds-safe interfaces).   I added an additional check for this case (the code for checking bounds declarations at calls already had a similar check).

Testing:
- Passed local testing on Windows x64.
- Passed Checked C, clang regression tests and LNT regression tests on Linux x64.
- We should look at further testing of other syntactic forms that can have interop types (variable
  assignments and calls), as well as testing of dynamic casts involving ptr source/destinations.