Ignore value preserving operations in CanonBounds.
We were ignoring value preserving operations in SemaBounds.cpp.  Move
this code to CanonBounds instead, so this is during when checking
expressions for syntactic equality, modulo equality information.
- In CanonBounds, check type information on operands for equality (ignoring
checkedness).  Two expressions might produce the same value, but have
different types, causing the meaning of the operator that uses them to change.
- We ignore checkedness for now because of issues around comparing
expressions built from variables with bounds-safe interfaces (the types
compare unequal, even though the meaning of most operations hasn't changed).
We'll revisit operations where checkedness affects the semantics later.

Broaden the set of rhs expressions for which we'll track equality
information to include all non-modifying expressions.  We still need to
handle the case where the lhs of an assignment is used by the rhs bounds.
This could happen directly using a variable name or indirectly using an
alias.

Fix up tests for which the compiler is now issuing bounds declaration
warnings.