Fix compiler crash rewriting return function pointer type.
The compiler hit an assert while rewrting types with bounds-safe interfaces
to make them checked types.   This problem was found by trying to call
the signal functionin  a checked scope.   The problem occurred when a
function return type was a function pointer type with a bounds-safe interface.
The construction of new type location information asserted about a mismatch
in the expected size of the type location buffer.

The problem was that we were rewriting parameter and return types and then
modifying types using interop type information.  We need to do this in the
opposite order to construct new type location information properly.

Testing:
- Add tests in the Checked C repo of calls in checked scopes to functions that
 return pointer types with bounds-safe interfaces.