Avoid spurious warnings during checking of bounds declarations. (#486)
We are doing type-preserving substitutions, so we don't expect semantic errors during substitution into bounds. There could be warnings because the compiler runs semantic analysis when a substitution happens.   This is problematic because there's no way to fix the warnings, the warnings could confuse users, and the warnings could also be escalated to errors, which would cause compilation failures.

This updates the compiler to suppress diagnostic messages during substitution of expressions.   We add a member on Sema to control whether diagnostic messages are being suppressed, and use an RAII pattern to set/restore this variable.   This is a vastly simplified version of the SFINAE (substitution failure is not an error) support code for C++ templates.   I thought about re-using the SFINAE code, but it did not seem worth the risk/trouble.

Testing:
- Added 2 new regression test cases.  A recent compiler change made the compiler stop showing the original reported warning (the compiler now recognizes that null implies that any parameter bounds is true, so when an actual argument is null, it doesn't even substitute into bounds) .  I had to write a more convoluted case to trigger the problem.
- Passes local testing on Windows.
- Passed automated testing on Linux.