Fix compiler crash reported by user. (#488)
This fixes issue #487, a debug compiler crash reported by a user.  In
the last change to SemaBounds.cpp, a call to
MakeAssignmentImplicitCastExplicit was removed from the code that
substitutes argument expressions into parameter bounds expressions.
This led to a subtle problem.  The substitution code uses
lib/Sema/TreeTransform.  When an argument expression was an implicit
cast expression, TreeTransform modified it.  The argument expression
is used elsewhere in the AST, so this is incorrect.

TreeTransform assumes that implicit casts are being regenerated by
semantic analysis and that it is OK to modify them.  During
substitution, it rebuilds the bounds expression.  This calls semantic
analysis, which may call ImpCastExprToType. This implicitly casts an
expression to a new type.  If the expression is already an implicit
cast expression, it modifies the expression. The fix is to convert the
argument expression to an explicit cast expression when it is
substituted.

Testing:
- Added two regression cases: the user-reported case and a
cut-down version of the user code.