build release compilers for Windows (#495)
This fixes the build automation scripts for Windows to actually build release versions of clang.  We were relying on CMAKE_BUILD_TYPE, which is disregarded when cmake generates a build system for Visual Studio.  The fix is to pass the build type in as an additional parameter to the msbuild invocations instead.

This addresses issue #494, where the installed compiler required a debug version of the Windows CRT that isn't usually installed on Windows systems.

Testing:
- Build x86/x64 compiler installers.