Expressions that modify variables used in bounds declarations. (#503)
We are currently checking that bounds declarations are valid after assignments to variables with bounds declarations.  We also need to check that bounds declarations are valid after assignments to variables _used_ in bounds declarations.   This work is described in issue #502.

This change implements the first part of the work in issue #502.   It computes a mapping from expressions that modify variables to sets of in-scope bounds declarations that use the modified variables.

Determining whether an expression modifies a variable used by an in-scope bounds declaration is a lexically-scoped computation     When checking bounds declarations, we are currently traversing the AST in a lexically-scoped fashion.  We plan to change the traversal of the AST to be based on control flow, so that we can incorporate dataflow facts.   It isn't easy in clang to determine the lexical scope that an expression occurs in.   Clang doesn't have back pointers from expressions to their parent expressions or from statements to enclosing lexical scopes.  That makes it difficult to answer the question of what set of bounds declarations are in scope.  To deal with this, we compute a map that can be consulted at any point in a function.

We compute the information for each function, but do not use it.  For compilation units that don't use Checked C extensions, the maps will be empty.  We want to get coverage on the different kinds of ASTs that we might encounter, so the code is left on during bounds declaration checking.

Testing:
- I wrote a new set of Checked C clang tests that contain expressions that modify variables used in bounds declarations.  I then used this to dump the results of the analysis and manually inspect that the analysis is producing expected results.  The compiler will eventually produce error messages in some cases.   These are in tests/CheckedC/static-checking/bounds-side-effects.c.
- Passed automated testing on Windows and Linux.