Checkpoint work on using CFG during bounds declaration checking.
Instead of traversing the AST during bounds declaration checking, we
are building a control-flow graph and traversing that.  This is in
preparation for incorporating dataflow information during bounds
declaration checking.

The main complication is that some subexpressions are given top-level
CFG elements.  We have to avoid traversing them twice.

Passes 58/70 tests.