Work around clang CFG construction bug.
Clang produces a bad CFG for an increment expression in a for loop
that has an expression statement (a GCC extension) as the increment
expression and where the increment expression has a continue
statement:

void pr8880_3(int first) {
 for (int i  = 0; i > 0; (void)({ if (i > 0) { i--; continue; } 0; })) {}
}

The continue is supposed to go to the for-loop increment expression
in the CFG, but the target of the continue statement hasn't been set up
yet.  The lack of a target for the continue statement is detected and
the CFG is marked as bad.  However, an assert later fires.  The assert
is followed by a check for a bad CFG.  The workaround is to move the assert
after the check.  That way we don't assert a property that may not be
true for an ill-formed CFG.