Add constraints for checked c array type (#527)
When constraints for existing checked arrays were encountered, we threw an exception because we didn't know how to support them. I'm pretty sure the way to support them is to constrain the appropriate constraint variables to "arr" which will then be untouched by the rest of the re-writer right now. T

This adds code to do that and a test case which crashed before and doesn't crash now.

Fixes #519