Header Inclusion Update script (#529)
This PR contains a script update-headers.py which checks a list of .c or .h files for inclusions of standard c headers that have checkedC replacements and overwrites those to include the checkedC _checked.h versions instead. Script requires a file containing a list of filenames (one per line, but the line may contain some trailing characters such as \ or whitespace). Script will also take an -includeDir option to specify where to look for the _checked.h headers. Default location assumes that both the utils and checkedc repo are part of a llvm+checkedc-clang repo.

This PR also:
* Renames run.py to convert-commands.py.
* Adds a README.md to the directory describing both Python utilities.