During automation, set up external script for signing installer. (#538)
This updates the Windows automation scripts to optionally download a repo containing external scripts for signing the LLVM installer.   The external scripts are Microsoft-specific and we keep them in a private git repo in Visual Studio Online.   We clone the repo to automation/Windows/sign.

The tricky part about this was authenticating git commands involving VSO access.   VSO has an OAuth token that you can provide to build phases that can be passed through git using an extra http header.   In our automated builds, we add a separate VSO build task for signing an installer once one has been built.