Add support for _Return_value expressions (#544)
A_Return_value expression allows programmers to name the value returned by a function in a return bounds declaration.  This changes extends the clang IR to represent _Return_value expressions and _Current_expr_value expressions. It also adds lexing, parsing, and type system support for _Return_value expressions.   This addresses the first part of issue #205.

We create a new expression class BoundsValueExpr to represent these expressions.  This class is modelled on CXXThisExpr.   Constructing a_Return_value expression requires knowing the return type of the function that the return bounds is associated with.  We must have the _Return_value expression constructed before we can build the function type, since function types incorporate the return bounds expressions.  Unfortunately, we don't know the function return type until the middle of GetFullTypeForDeclarator in SemaType.cpp.  Because clang intertwines AST building, type checking, and other static semantic checking, we have to "defer parse" return bounds expressions.   "Defer parse" means that the compiler recognize the tokens that make up a return bounds expression, saves them, and parses them later to build the IR for the return bounds expressions.

To allow parsing of the return bounds to happen in the middle of SemaType.cpp, we introduce a callback from semantic checking to the parser.  This lets us preserve the interface between the two.  We have to bring the parameters into scope, since they've already gone out of scope by the time GetFullTypeForDeclarator runs.   We aren't the first people to have to do this.  Other parsing phases such as parsing of attributes have to do this too.

After parsing the return bounds expression, we attach it to the declarator whose type is being constructed.    This lets us get the line number information right for error messages involving redeclarations with conflicting return bounds.  (Long explanation: memoization of function types also memoizes bounds expressions embedded in types.  It does this by ignoring line numbers and abstracting function parameters to parameter locations.   The end result is that line numbers for expressions embedded in types are meaningless.  Clang has a complicated mechanism for tracking source line information for types called TypeSourceInfo, but we haven't implemented support for embedded expressions in types and are not likely to anytime soon.

This change also contains some unrelated clean up:
1. Fix the line ending for the test case for bug526 to be line feeds.
2. Delete identifiers for _Dynamic_bounds_cast and _Assume_bounds_cast.  These are keywords, so we don't need identifiers.

Testing:
-  There will be a separate pull request to the Checked C repo for parsing and typechecking tests for_Return_value.
- Added clang tests for AST dumping and checking of bounds declarations.  The checking of bounds declarations involving _Return_value doesn't work yet, so the tests document this.
- Passes automated Checked C and clang testing with these changes.