Fix errors related to returning a CHKCBindTemporaryExprClass.
There are lots of places in the compiler that do structural matching on
a StringLiteral expression being present at a specific point in the AST.
We're now returning a CHKCBindTemporaryExprClass instead.  This changes
patches up a number of a places to skip the expression temporary.  There
are still about 26 failing Clang regression tests, though.