Fix duplicate messages for void pointer arithmetic.
The compiler was emitting duplicate error messages for void pointer
arithmetic involving checking pointers in bounds expressions.  The problem
was that the logic for marking the expression is invalid is separate from
actually emitting the diagnostic.  The logic only looked for the language
being C++.  The fix is to look for the conditions related to Checked C also.

This caused a duplicate error messages beccause the expression wasn't turned
into an invalid expression.  When a subsequent TreeTransform pass for bounds
declaration checking ran, it emitted the diagnostic a second time when it
encountered the expression again.

I noticed that we had some logic for handling pointer arithmetic involving
function pointers.  This is allowed by a GCC extension.  For checked pointers,
we never allow _Array_ptrs to function pointers.  We only allow _Ptrs to
function pointers.  Pointer arithmetic is never allowed for _Ptrs, so we
never reach this logic.  I converted the logic to an assert and added
some comments.

Testing:
- I added a new Checked C clang regression test for these cases that checks
 that only   one error message is emitted.