Fix duplicate messages for illegal void pointer arithmetic. (#563)
The compiler was emitting duplicate error messages for void pointer arithmetic using checked pointers in bounds expressions.  The problem was that the logic for marking the expression as invalid is separate from actually emitting the diagnostic.  The logic only looked for the language being C++.  The fix is to look for conditions related to Checked C also.  This fixes issue #543.

This caused a duplicate error messages because the expression wasn't turned into an invalid expression.  When a subsequent TreeTransform pass for bounds declaration checking ran, it emitted the diagnostic a second time when it encountered the expression again.

I noticed that we had some logic for handling pointer arithmetic involving checked function pointers.  This is allowed by a GCC extension.  For checked pointers, we never allow _Array_ptrs to function pointers.  We only allow _Ptrs to function pointers.  Pointer arithmetic is disallowed for _Ptrs, so we don't need a message specific to checked function pointers.   I removed the logic and added some comments.

Testing:
- I added a new Checked C clang regression test for these cases that checks hat only   one error message is emitted.
- Passed automated testing.