Update semantic checking for checked scopes.
This change updates semantic checking for checked scopes to no longer
rely on the clang "scope" class.  The scope class is meant to be used for
identifier lookup during parsing, not to represent properties of  a scope
during semantic checking.

We've been using the "scope" class, but it isn't the right approach.  We
end up using getCurScope(), which is only valid during parsing.  The code
specifically warns against using it during semantic checking.  Use of it
could lead to problems when re-running semantic checking.   The implementation
also has some layering violations, where parsing is doing things that are
semantic checking (and inferring whether a scope is checked or not).

We use two approaches to fix this:
- We annotate the checked scope informatin in the source directly on the AST.
- During semantic checking, we track the checked scope state, updating it
  based on information in the AST.

The implementation isn't quite done.  There is an issue with the checked
state inforrmation not being set up quite right when checking function
declarations.