Improve IR and implementation of checked scopes. (#567)
This change updates the IR and semantic checking for  checked scopes, 
in preparation for allowing multiple kinds of checked scopes.     We
were using the clang "scope" class.  The scope class is meant to be used for
identifier lookup during parsing, not to represent properties of  a scope
during semantic checking.

When parsing declarations, _Checked is now treated as declaration 
specifier, unless it is followed by `{` or `[`.  It is only allowed on function
declarations.

In the IR, we introduce an enum for checked specifier information and
another enum for a  checked scope modifier.  We record the checked specifier
information in the source in the AST for compound statements.  We also
record the checked scope state, which combines inherited information with
any checked scope specifier on the compound statement.

The clang "scope" class wasn't the right place to store the information.   We
ended up using getCurScope(), which is only valid during parsing.  The code
specifically warns against using it during semantic checking.  Use of it
could lead to problems when re-running semantic checking.   The implementation
also had some layering violations, where parsing is doing things that are
semantic checking (and inferring whether a scope is checked or not).

During semantic checking, we now tracked the checked scope state, updating
it based on information in the AST.

Testing:
- Passed automating testing on Linux for x64 debug and release.
- Passed local testing on Windows.